+++ apikeyrotator/__init__.py
@@ -98,7 +98,7 @@ from .utils import (
 )
 
 
-__version__ = "0.9.0"
+__version__ = "0.9.1"
 __author__ = "Prime Evolution"
 
 __all__ = [